Python3 backport

+++ tests/func/run_tests_docker.sh
@@ -23,6 +23,7 @@ teardown() {
 trap teardown EXIT INT TERM
 
 install_rpm=${TBD_INSTALL_RPM:-0}
+export PYTHON=${PYTHON-python2}
 
 # For circle-ci tests we want to install using RPM
 # When launched locally we install via pip
@@ -35,19 +36,32 @@ then
     yum -d1 "--disablerepo=pgdg*"  install -y $rpm
     rpm --query --queryformat= temboard-agent
 else
-    pip2 install -e .
-    if type -p yum ; then
+    $PYTHON -m pip install -e .
+    if type -p yum &>/dev/null && $PYTHON --version | grep -F 'Python 2' ; then
 	    yum -q -y "--disablerepo=pgdg*" install python-psycopg2
     else
-	    pip2 install psycopg2-binary
+	    $PYTHON -m pip install psycopg2-binary
     fi
 fi
 
-pip2 install pytest
+$PYTHON -m pip install pytest pytest-mock
+
+for locale in C.UTF-8 en_US.utf8 ; do
+	if locale -a | grep -q "$locale" ; then
+		export LC_ALL="$locale"
+		break
+	fi
+done
 
 export TBD_PGBIN=$(readlink -e /usr/pgsql-${TBD_PGVERSION}/bin /usr/lib/postgresql/${TBD_PGVERSION}/bin)
+export TBD_WORKPATH="/tmp"
 
 # Remove any .pyc file to avoid errors with pytest and cache
 find . -name \*.pyc -delete
 rm -rf /tmp/tests_temboard
-make -C tests/func/ pytest PYTEST_ARGS="${PYTEST_ARGS-$*}"
+sudo -Eu testuser \
+	/usr/bin/env PATH="$PATH" \
+	"$PYTHON" -m pytest \
+	-vv --capture=no -p no:cacheprovider \
+	tests/func/ \
+	"$@"
